recommendation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-surprise==1.1.3
sentence-transformers==2.2.2
apscheduler==3.10.4
prometheus-client==0.19.0
python-json-logger==2.0.7
httpx==0.25.2
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2
pytest==7.4.3
